change to (u32 *)(Data+nvram_pos) in src/cpu/amd/agesa/s3_resume.c
As Paul pointed in AMD S3 resume: Add framwork to write bigger data[1],
change pointer from (u8 *)(Data+nvram_pos) to (u32 *)(Data+nvram_pos).
I have tested on Parmer.

[1] http://review.coreboot.org/#/c/3413/

+++ src/cpu/amd/agesa/s3_resume.c
@@ -272,9 +272,9 @@ u32 OemAgesaSaveS3Info(S3_DATA_TYPE S3DataType, u32 DataSize, void *Data)
 
 	for (nvram_pos = 0; nvram_pos < DataSize - CONFIG_AMD_SB_SPI_TX_LEN; nvram_pos += CONFIG_AMD_SB_SPI_TX_LEN) {
 		data = *(u32 *) (Data + nvram_pos);
-		flash->write(flash, nvram_pos + pos + 4, CONFIG_AMD_SB_SPI_TX_LEN, (u8 *)(Data + nvram_pos));
+		flash->write(flash, nvram_pos + pos + 4, CONFIG_AMD_SB_SPI_TX_LEN, (u32 *)(Data + nvram_pos));
 	}
-	flash->write(flash, nvram_pos + pos + 4, DataSize % CONFIG_AMD_SB_SPI_TX_LEN, (u8 *)(Data + nvram_pos));
+	flash->write(flash, nvram_pos + pos + 4, DataSize % CONFIG_AMD_SB_SPI_TX_LEN, (u32 *)(Data + nvram_pos));
 
 	flash->spi->rw = SPI_WRITE_FLAG;
 	spi_release_bus(flash->spi);
